requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
envparse==0.2.0
psycopg2==2.7.7
pytz==2018.9
